+++ lib/widget.dart
@@ -1,1 +1,0 @@
-export 'widget/extension.dart';